requirements.txt
python-dotenv==0.10.1
pytz==2018.7
pexpect==4.6.0